2.0V 线条修改

+++ Uni/Model/UNIShopModel.h
@@ -15,8 +15,8 @@
 @property(nonatomic,copy)NSString* address;  //店铺地址
 @property(nonatomic,copy)NSString* telphone;  //店铺电话
 @property(nonatomic,assign)int shopId;   //店铺ID
-@property(nonatomic,assign)float x;      //坐标X
-@property(nonatomic,assign)float y;      //坐标Y
+@property(nonatomic,assign)double x;      //坐标X
+@property(nonatomic,assign)double y;      //坐标Y
 
 -(id)initWithDic:(NSDictionary*)dic;
 @end
